requirements.txt
fastapi>=0.120
uvicorn>=0.30
pydantic>=2.8
pydantic-settings>=2.2
python-dotenv>=1.0
PyYAML>=6.0
typer>=0.12
rich>=13.7
jsonschema>=4.22
google-genai>=0.6

